Added option to reset window close flag
This should(mandatory) called in draw functions. This reset function
basically checks if window was asked to be closed by the user earlier
if so, it resets the flag and raises the window to focus.

This feature basically enables reuse same window multiple times
in the same program and helps avoid creating uncessary additional
windows. However for this feature to work as expected, the
fg::Window::close function should be called in the following loop
constructs only.
* `do-while`
* `for`
`while` cannot work as the condition check is made prior to any draw
call where the window close flag check is happening.